data-vis/requirements.txt
gradio==5.50.0
rpi-lgpio==0.6
smbus2==0.4.3